Add files via upload
Z algorithm is used to find the occurrence of a pattern in a string in linear time. Suppose if the length of the string is n and the size of the pattern to be searched is m, the time taken to solve will be of the order O(m+n). The z-algorithm uses a Z array to find the occurrence of a pattern.