Add files via upload
This is my final year project which aim to use machine learning algorithm to predict stock market. It first use tsne to find similar dataset with target dataset and then create a LSTM model and trained with similar dataset which found in order to predict target ticker accurately. It also support few shot learning as it does not require a large amount of training dataset.